version 1.7 repair bugs

+++ auto_messages.js
@@ -1,8 +1,7 @@
 const { MessageEmbed } = require('discord.js');
-const displayVoiceStats = require("./displayVoiceStats.js");
+const displayVoiceStats = require("./voiceStats/displayVoiceStats.js");
 
-exports.auto_messages = (client, date, database, gif, serverId, channelDateId, channel_list, statistics) => {
-
+module.exports = (client, date, database, gif, serverId, channelIDs, channelNames, channelNameStatisticsFunctions, checkPremissions) => {
 
   setInterval(() => {
 
@@ -11,55 +10,68 @@ exports.auto_messages = (client, date, database, gif, serverId, channelDateId, c
 
     if (minute == 0) {
       //wysyłanie godziny na kanał
-      var channel = client.channels.cache.find(channel => channel.name === channel_list[1]);
-      channel.send("Jest godzina " + date.hour() + ":00");
+      const channel = client.channels.cache.find(channel => channel.name === channelNames.hourChannel);
+      if (checkPremissions(channel, channelNames.hourChannel))
+        channel.send("Jest godzina " + date.hour() + ":00");
     }
+
     //statystyki, ustawianie nowej daty
     if (hour == 0 && minute == 0) {
-      statistics.new_date(client, serverId, channelDateId);
+      channelNameStatisticsFunctions.new_date(client, serverId, channelIDs.channelDateId);
     }
 
     //wysylanie wiadomosci codziennej
     if (hour == 10 && minute == 0) {
-      var channel = client.channels.cache.find(channel => channel.name === channel_list[8]);
-      const embed = new MessageEmbed()
-        .setColor('#0099ff')
-        // .setTitle('Codzienna wiadomość')
-        .setTitle(`Dzisiaj mamy   📅   ${date.day_of_the_week()}, ${date.full_day_message()}`)
-        .setDescription('Miłego dnia 💚')
-        .setAuthor('Dzień dobry 🖐')
-        .setTimestamp()
-      channel.send({ embeds: [embed] });
+      const channel = client.channels.cache.find(channel => channel.name === channelNames.botChannel);
+      if (checkPremissions(channel, channelNames.botChannel)) {
+        const embed = new MessageEmbed()
+          .setColor('#0099ff')
+          // .setTitle('Codzienna wiadomość')
+          .setTitle(`Dzisiaj mamy   📅   ${date.day_of_the_week()}, ${date.full_day_message()}`)
+          .setDescription('Miłego dnia 💚')
+          .setAuthor('Dzień dobry 🖐')
+          .setTimestamp()
+        channel.send({ embeds: [embed] });
+      }
+
+
     }
 
-    //wysylanie losowej wiadomosci na kanal wiewiorka
+    // wysylanie losowej wiadomosci na kanal wiewiorka
     if ((hour == 11 && minute == 0) || (hour == 22 && minute == 0)) {
-      var channel = client.channels.cache.find(channel => channel.name === channel_list[8]);
-      (async () => {
-        channel.send( await database.rand_message("DIFFERENT_MESSAGES"));
-      })();
+      const channel = client.channels.cache.find(channel => channel.name === channelNames.botChannel);
+      if (checkPremissions(channel, channelNames.botChannel)) {
+        (async () => {
+          channel.send(await database.rand_message("DIFFERENT_MESSAGES"));
+        })();
+      }
     }
 
     if (minute == 0 || minute == 10 || minute == 20 || minute == 30 || minute == 40 || minute == 50) {
       (async () => {
-        var channel = client.channels.cache.find(channel => channel.name === channel_list[11]);
-        let messages = await channel.messages.fetch();
-        messages.forEach(msg => {
-          msg.delete();
-        })
-        await displayVoiceStats.send_time_voice(channel);
+        const channel = client.channels.cache.find(channel => channel.name === channelNames.voiceStatisticsChannel);
+        if (checkPremissions(channel, channelNames.voiceStatisticsChannel)) {
+          let messages = await channel.messages.fetch();
+          messages.forEach(msg => {
+            msg.delete();
+          })
+          await displayVoiceStats.send_time_voice(channel);
+        }
       })();
     }
 
 
-    //wysyłanie gifow
+    //   //wysyłanie gifow
     if ((minute == 0 || minute == 20 || minute == 40)) {
-      var channel = client.channels.cache.find(channel => channel.name === channel_list[2]);
-      (async () => {
-        channel.send(await gif.tenor_gif(await gif.rand_gif_category()));
+      const channel = client.channels.cache.find(channel => channel.name === channelNames.gifsChannel);
+      if (checkPremissions(channel, channelNames.gifsChannel)) {
+        (async () => {
+          channel.send(await gif.tenor_gif(await gif.rand_gif_category()));
+        }
+        )();
       }
-      )();
     }
+
   }, 59000);
 
 }
